refactor(notification) : 예외 클래스 위치 변경
