+++ src/test/java/com/codeit/mopl/domain/notification/service/NotificationServiceUnitTest.java
@@ -16,8 +16,8 @@ import com.codeit.mopl.domain.notification.entity.Notification;
 import com.codeit.mopl.domain.notification.entity.SortBy;
 import com.codeit.mopl.domain.notification.entity.SortDirection;
 import com.codeit.mopl.domain.notification.entity.Status;
-import com.codeit.mopl.domain.notification.exception.NotificationForbidden;
-import com.codeit.mopl.domain.notification.exception.NotificationNotFoundException;
+import com.codeit.mopl.exception.notification.NotificationForbidden;
+import com.codeit.mopl.exception.notification.NotificationNotFoundException;
 import com.codeit.mopl.domain.notification.mapper.NotificationMapper;
 import com.codeit.mopl.domain.notification.repository.NotificationRepository;
 import com.codeit.mopl.domain.user.entity.User;
